requirements.txt
langchain-core
langchain-community
langchain-google-genai
youtube-transcript-api
python-dotenv